+++ src/utils/macros.rs
@@ -42,11 +42,15 @@ macro_rules! load_manifest {
 macro_rules! write_manifest {
     ($var:expr => $path:expr) => {
         let tomlfy: String = toml::to_string_pretty($var)?;
-        std::fs::write($path, tomlfy)?
+        if !crate::utils::dryrun::get_dry_run() {
+            std::fs::write($path, tomlfy)?
+        }
     };
     ($var:expr) => {
         let tomlfy: String = toml::to_string_pretty($var)?;
-        std::fs::write("./typst.toml", tomlfy)?
+        if !crate::utils::dryrun::get_dry_run() {
+            std::fs::write("./typst.toml", tomlfy)?
+        }
     };
 }
